+++ src/py42/_internal/token_providers.py
@@ -77,6 +77,7 @@ class C42APILoginTokenProvider(C42APITmpAuthProvider):
         response = self._auth_session.post(uri, data=json.dumps(data))
         return response
 
+
 class C42APIStorageAuthTokenProvider(C42APITmpAuthProvider):
     def __init__(self, auth_session, plan_uid, destination_guid):
         super(C42APIStorageAuthTokenProvider, self).__init__()