+++ src/py42/modules/securitydata.py
@@ -2,10 +2,7 @@ import json
 from threading import Lock
 from requests.exceptions import HTTPError
 
-from py42.sdk.exceptions import (
-    Py42SecurityPlanConnectionError,
-    raise_py42_error,
-)
+from py42.sdk.exceptions import Py42SecurityPlanConnectionError, raise_py42_error
 
 
 class SecurityModule(object):